+++ spec/requests/waterfalls_spec.rb
@@ -23,7 +23,7 @@ RSpec.describe "Waterfalls", type: :request do
 
       expect(response).to have_http_status(:ok)
       expect(response.body).to include('data-controller="explore-map"')
-      expect(response.body).to include("Sign in")
+      expect(response.body).to include(I18n.t("layouts.header.sign_in"))
       expect(response.body).to include("Sekumpul Waterfall")
     end